+++ UniVerseAPI/universe/src/main/java/com/unidev/universe/entities/Like.java
@@ -13,7 +13,7 @@ public class Like {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "like_id")
-    private int likeId;
+    private Long likeId;
 
     @NotNull
     @ManyToOne
